refactor folders and project structure
