+++ lib/lib/WsStream.js
