+++ lib/lib/decoder.js
