+++ lib/lib/network-registry.js
