+++ lib/middleware/index.js
