+++ lib/register-websockets.js
