+++ package.json
@@ -1,17 +1,18 @@
 {
   "name": "tunnel-server",
-  "version": "0.3.2",
+  "version": "0.3.1",
   "scripts": {
-    "clean": "rm -rf bin/server",
+    "clean": "rm -rf lib",
     "start-dev": "node index.js",
-    "lint": "eslint *.js server/",
-    "compile": "babel server --out-dir bin/server",
+    "lint": "eslint *.js src/",
+    "compile": "babel src --out-dir lib",
     "build": "npm run lint && npm run clean && npm run compile",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "bin": {
     "tunnel-server": "bin/index.js"
   },
+  "main": "lib/index.js",
   "dependencies": {
     "babel": "^5.8.21",
     "body-parser": "^1.13.3",
