+++ src/lib/WsStream.js
