+++ src/lib/decoder.js
