+++ src/lib/network-registry.js
